+++ mobile/views/Home/HomeView.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { View, Alert } from "react-native";
 import { useSetRecoilState } from "recoil";
-import ClearAsyncStorageButton from "../components/ClearAsyncStorageButton";
-import LockButton from "../components/LockButton";
-import { LockLoadingState, LockState } from "../recoil_state";
+import ClearAsyncStorageButton from "../../components/ClearAsyncStorageButton";
+import LockButton from "../../components/LockButton";
+import { LockLoadingState, LockState } from "../../recoil_state";
 import { WS_URL } from "@env";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { sendPostRequest } from "../helpers/Requests";
+import { sendPostRequest } from "../../helpers/Requests";
 
 interface WebSocketMessage {
   topic: string;
