--- mobile/views/PlaceholderView.tsx
@@ -1,3 +1,0 @@
-export default function PlaceholderView() {
-    return <></>
-}
